backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.13.0
email-validator==2.3.0
sqlalchemy==2.0.47
# Driver database (async)
asyncpg==0.31.0
asyncmy==0.2.10
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
redis==7.2.1
orjson==3.11.5
starlette==0.52.1
bleach==6.2.0
tinycss2==1.4.0
sse-starlette==3.3.2
python-multipart==0.0.20
Pillow==11.1.0
